requirements.txt

fastapi
uvicorn
tensorflow
tensorflow-hub
pillow
numpy
python-multipart